Move extract_raw_data_from_model() and insert_raw_data_into_model() methods
